src/server/pyproject.toml
[project]
name = "server"
description = "None"
readme = "README.md"
version = "0.1.0"

requires-python = ">=3.13"

dependencies = [
    "django>=5.2.4",
    "django-htmx>=1.23.2",
    "django-vite>=3.1.0",
    "django-cotton>=0.2.0",
    "sqids>=0.5.2",
    "python-dotenv>=1.1.1",
    "django-widget-tweaks>=1.5.0",
    "django-unfold>=0.65.0",
    "django-ninja>=1.3.0",
    "email-validator>=2.0.0",
    "django-allauth>=65.3.0",
    "django-file-form>=4.0.0",
    "boto3>=1.35.0",
    "django-storages>=1.14.0",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "factory-boy>=3.3.3",
    "honcho>=2.0.0",
]

[tool.djlint]
ignore = ["D018"]
custom_html = "c-[\\w-]+"
